+++ client/index.html
@@ -12,6 +12,7 @@
     <!-- tabular  https://tabulator.info/ -->
     <link href="https://unpkg.com/tabulator-tables@6.1/dist/css/tabulator.min.css" rel="stylesheet">
     <link href="https://unpkg.com/tabulator-tables@6.1/dist/css/tabulator_bootstrap5.min.css" rel="stylesheet">
+    <!-- SPA entry -->
   </head>
   <body>
     <div id="app"></div>
@@ -20,10 +21,6 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
     <script type="text/javascript" src="https://unpkg.com/tabulator-tables@6.1/dist/js/tabulator.min.js"></script>
 
-    <!-- internal scripts -->
-    <script type="text/javascript" src="/static/index.js"></script>
-
-    <!-- SPA entry -->
     <script type="module" src="/src/main.jsx"></script>
 
   </body>
